refactor: extract Matching service into Action classes
Decompose MatchingEngine into MatchBranch, MatchAllUnmatched, and
RematchBranch actions under Domain/Matching/Actions. Update RunSyncJob
and SyncService to use new actions. Delete MatchingEngine, its interface,
and MatchingServiceProvider.
