+++ backend/bootstrap/providers.php
@@ -5,7 +5,6 @@ use App\Providers\Bitrix24ServiceProvider;
 use App\Providers\GitLabServiceProvider;
 use App\Providers\InboxServiceProvider;
 use App\Providers\LlmServiceProvider;
-use App\Providers\MatchingServiceProvider;
 use App\Providers\NarrativeServiceProvider;
 use App\Providers\ReportServiceProvider;
 use App\Providers\SyncServiceProvider;
@@ -16,7 +15,6 @@ return [
     GitLabServiceProvider::class,
     InboxServiceProvider::class,
     LlmServiceProvider::class,
-    MatchingServiceProvider::class,
     NarrativeServiceProvider::class,
     ReportServiceProvider::class,
     SyncServiceProvider::class,
